skip sequence
first idea to automate the skip using and array that is periodicaly
banged a random 0 or 1 to mark a skip in the time line. quite extreme
at the moment, needs to be more subtle. also added automation to the
masking, when patch loads, a random track position is chosen and makes
a mask at that point