Xcode handling the storage of play_cont - but there is a bug in the pd patch
The pd patch receives the play_cont from xcode in the sequence
sub-patch, with the receive ‘r fibCountFromXcode’. It successfully
writes that to the playcont table, but I don’t think it updates the
rest of the patch properly.
You will need to replicate whatever logic occurs when the loadbang
happens and you read play_cont from the file, so that it also occurs
when the r fibCountFromXcode receives a new value.
One of the tricks I’ve done with Pd before is replace all my loadBangs
with r my_loadbang, and that gives me a quick and easy way to
re-initialise the entire patch with re-opening it. Doing that, then
when you receive fibCountFromXCode, you can forceably reinit the entire
patch and the whole thing should work.